Make sure dataflow server uses h2 launched by spring cloud
If user runs `spring cloud dataflow h2` you'd expect the database to
be used by dataflow. It's tricky to achieve, and impossible with
an in-memory data store (because of the way dataflow is set up
currently). This was achieved by adding a feature: each deployable
can specify "enabled" (in addition to existing "disabled" feature)
properties: i.e. environemnt properties that are applied to all
other apps as long as this one is running.

Another change here is to pass the "PATH" env var from the deployer
down to deployed apps. This affects dataflow in particular because
it likes to launch other apps, and it will pick the wrong Java
version (for instance) if the path is not preserved.