Add generic deployment properties to deployer apps
Instead of JAVA_OPTS (didn't like the name anyway), user can define
generic "properties" in each deployable that will be passed down
to the app request.

The current 1.0 snapshots of the local deployer look in there for
JAVA_OPTS (but that hopefully will change to something more
"property like") before 1.0.3 is released.