Replace static version field with ClassPathResource.
Previously, the default version was loaded from a java constant. This version used to be updated by hand after the release train was published to maven central, then built and deployed by hand. Spring-Cloud-Cli is now built and deployed automatically, but the default version hasn't been updated. The default version was only used if the version couldn't be read from the manifest (which should be the version used during release.

The constant was replaced with the contents of a file that is updated at build time. This process works for development and release without any manual intervention.

Fixes gh-174